`push`: Support creation of new pages and pruning of remote pages
Pages can be created with or without parent. Both the category and
parent are implied from the `.md` file's path.

When specifying the `--prune` option for the `push` command, remote
pages that do not exist locally will be deleted.

Also made a few other refactorings, mostly related to searching /
filtering pages in a Catalog instance.