Add support for Content Filters
Content Filters allow for a different representation of content on
Readme.io than in local `.md` files.

Filters have a chance to alter content on the way to Readme.io and on the
way back. The changes they perform should be symmetrical: pushing a page
to Readme with a filter enabled followed by fetching that same page from
Readme should result in the exact same local content.

The first content filter, `hostedFiles` allows for hosting images on a
public Web server. Relative image paths are transformed to their
equivalent public URL and transformed back to their relative counterpart
on the way back.