Make the `hostedFiles` filter consider both links and images
This allows for links to content files other than images (ex: PDF files).

Also, tweaked the algorithm to perform replacements in page content.

The previous strategy didn't take into account that replacing content
often changes its length and thus invalidates the element positions that
are pre-calculated on Page creation.

Batching the replacements in a single operation and starting from the
element that is the farthest in the content addresses this.