Add support for custom page footers with the `footer` content filter
This filter allows to specify a Mustache template that will be rendered
just before pushing the content of pages to the Readme API.

Any configuration attribute defined for the filter in `config.yml` will
be passed to the Mustache rendering context to be used as dynamic
content.

The footer is removed from the content when fetching content from
Readme and will be updated on every page whenever the template changes.