Rework iOS calling sections
Use capitalization on call combination identifiers, e.g. App-to-SIP
instead of app-to-sip (it's especially important in the SIP case where
SIP is an acknowledged acronym).

Fix link on TTS (point to cloud, not on-premise)

Remove "or for video call" code snippet. That should be covered in
video specific section.

Remove old content on iOS Local Notifications (we will put emphasis on
VoIP push notifications and CallKit)

Add callout with reference to our CallKit sample app.

Add link from Setting up App-to-App to Handling Incoming Calls.